refactor(windows): centralize current-user identity lookup
